pyproject.toml
[project]
name = "vanilla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pystache>=0.6.7",
    "pyyaml>=6.0.3",
]
